ログイン状態の場合にのみ、自身が出品した商品情報を削除できることをコントローラー側でも条件を設ける

+++ app/controllers/items_controller.rb
@@ -35,8 +35,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-     redirect_to root_path
+    if current_user == @item.user
+      if @item.destroy 
+       redirect_to root_path
+      end
     end
   end
